--- buildSrc/src/main/kotlin/kotlin-multiplatform-conventions.gradle.kts
@@ -1,141 +1,0 @@
-import org.gradle.api.tasks.testing.logging.*
-import org.gradle.api.tasks.bundling.Jar
-import org.jetbrains.kotlin.gradle.dsl.*
-import org.jetbrains.kotlin.gradle.dsl.abi.ExperimentalAbiValidation
-import org.jetbrains.kotlin.gradle.targets.jvm.KotlinJvmTarget
-
-plugins {
-    kotlin("multiplatform")
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-kotlin {
-    if (abiCheckEnabled) {
-        @OptIn(ExperimentalAbiValidation::class)
-        abiValidation {
-            enabled = true
-        }
-    }
-
-    jvm {
-        compilations.all {
-            compileTaskProvider.configure {
-                compilerOptions.jvmTarget = JvmTarget.JVM_1_8
-            }
-        }
-    }
-    jvmToolchain(jdkToolchainVersion)
-    // According to https://kotlinlang.org/docs/native-target-support.html
-    // Tier 1
-    linuxX64()
-    macosArm64()
-    iosSimulatorArm64()
-    // Tier 2
-    linuxArm64()
-    watchosSimulatorArm64()
-    watchosArm32()
-    watchosArm64()
-    tvosSimulatorArm64()
-    tvosArm64()
-    iosArm64()
-    // Tier 3
-    androidNativeArm32()
-    androidNativeArm64()
-    androidNativeX86()
-    androidNativeX64()
-    iosX64()
-    mingwX64()
-    watchosDeviceArm64()
-    // Deprecated for removal: see KT-78660
-    @Suppress("DEPRECATION", "DEPRECATION_ERROR")
-    run {
-        macosX64()
-        tvosX64()
-        watchosX64()
-    }
-    js {
-        outputModuleName = project.name
-        nodejs()
-    }
-    @OptIn(org.jetbrains.kotlin.gradle.ExperimentalWasmDsl::class)
-    wasmJs {
-        // Module name should be different from the one from JS
-        // otherwise IC tasks that start clashing different modules with the same module name
-        outputModuleName = project.name + "Wasm"
-        nodejs()
-    }
-    @OptIn(org.jetbrains.kotlin.gradle.ExperimentalWasmDsl::class)
-    wasmWasi {
-        nodejs()
-        compilations.configureEach {
-            compileTaskProvider.configure {
-                compilerOptions {
-                    optIn.add("kotlin.wasm.internal.InternalWasmApi")
-                }
-            }
-        }
-    }
-    applyDefaultHierarchyTemplate()
-    sourceSets {
-        commonTest.dependencies {
-            implementation("org.jetbrains.kotlin:kotlin-test:${version("kotlin")}")
-        }
-        jvmMain.dependencies {
-            compileOnly("org.codehaus.mojo:animal-sniffer-annotations:1.20")
-            // Workaround until https://github.com/JetBrains/kotlin/pull/4999 is picked up
-            api("org.jetbrains:annotations:23.0.0")
-        }
-        jvmTest.dependencies {
-            implementation("org.jetbrains.kotlin:kotlin-test-junit:${version("kotlin")}")
-            implementation("junit:junit:${version("junit")}")
-        }
-        groupSourceSets("jsAndWasmShared", listOf("web", "wasmWasi"), listOf("common"))
-    }
-
-    @OptIn(org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi::class)
-    compilerOptions {
-        configureGlobalKotlinArgumentsAndOptIns()
-        freeCompilerArgs.add("-Xexpect-actual-classes")
-        optIn.add("kotlin.ExperimentalMultiplatform")
-    }
-}
-
-// Disable intermediate sourceSet compilation because we do not need js-wasm common artifact
-tasks.configureEach {
-    if (name == "compileJsAndWasmSharedMainKotlinMetadata") {
-        enabled = false
-    }
-    if (name == "compileJsAndWasmJsSharedMainKotlinMetadata") {
-        enabled = false
-    }
-}
-
-tasks.named("jvmTest", Test::class) {
-    testLogging {
-        showStandardStreams = true
-        events = setOf(TestLogEvent.PASSED, TestLogEvent.FAILED)
-    }
-    project.properties["stressTest"]?.let { systemProperty("stressTest", it) }
-}
-
-kotlin.targets.withType<KotlinJvmTarget>().configureEach {
-    // Fill attributes for the JVM implementation Jar only
-    tasks.named<Jar>(artifactsTaskName) {
-        fillManifestImplementationAttributes(project)
-    }
-}
-
-/*
- * To avoid a conflict with a JPMS module provided by kotlinx-coroutines-*-jvm artifacts,
- * an explicit automatic module name has to be specified in the manifest for metadata jars.
- */
-tasks.named("allMetadataJar", Jar::class) {
-    val moduleName =  project.name.replace("-", ".") + ".artifact_disambiguating_module"
-    manifest {
-        attributes("Automatic-Module-Name" to moduleName)
-    }
-}
