+++ buildSrc/src/main/kotlin/kotlin-shared-conventions.gradle.kts
@@ -1,9 +1,13 @@
 import org.gradle.api.tasks.testing.logging.TestLogEvent
 import org.gradle.kotlin.dsl.invoke
+import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.dsl.*
 import org.jetbrains.kotlin.gradle.dsl.abi.AbiValidationExtension
 import org.jetbrains.kotlin.gradle.dsl.abi.AbiValidationMultiplatformExtension
 import org.jetbrains.kotlin.gradle.dsl.abi.ExperimentalAbiValidation
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompilationTask
+import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
+import org.jetbrains.kotlin.gradle.tasks.KotlinNativeCompile
 import org.jetbrains.kotlin.gradle.targets.jvm.KotlinJvmTarget
 
 private fun KotlinCommonCompilerOptions.configureGlobalKotlinArgumentsAndOptIns() {
@@ -18,6 +22,8 @@ private fun KotlinCommonCompilerOptions.configureGlobalKotlinArgumentsAndOptIns(
         "kotlinx.coroutines.FlowPreview"
     )
 }
+
+apply(plugin = "org.jetbrains.kotlinx.atomicfu")
 
 extensions.configure<JavaPluginExtension> {
     sourceCompatibility = JavaVersion.VERSION_1_8
@@ -69,18 +75,14 @@ plugins.withId("org.jetbrains.kotlin.multiplatform") {
         // According to https://kotlinlang.org/docs/native-target-support.html
         // Tier 1
         linuxX64()
-        macosX64()
         macosArm64()
         iosSimulatorArm64()
-        iosX64()
         // Tier 2
         linuxArm64()
         watchosSimulatorArm64()
-        watchosX64()
         watchosArm32()
         watchosArm64()
         tvosSimulatorArm64()
-        tvosX64()
         tvosArm64()
         iosArm64()
         // Tier 3
@@ -88,8 +90,17 @@ plugins.withId("org.jetbrains.kotlin.multiplatform") {
         androidNativeArm64()
         androidNativeX86()
         androidNativeX64()
+        iosX64()
         mingwX64()
         watchosDeviceArm64()
+
+        // Deprecated for removal: see KT-78660
+        @Suppress("DEPRECATION", "DEPRECATION_ERROR")
+        run {
+            macosX64()
+            tvosX64()
+            watchosX64()
+        }
         js {
             outputModuleName = project.name
             nodejs()
@@ -184,4 +195,42 @@ tasks.withType<Test> {
 
 tasks.named("check") {
     dependsOn(tasks.named("checkLegacyAbi"))
+}
+
+tasks.withType<KotlinCompilationTask<*>>().configureEach {
+    val isMainTaskName = name.startsWith("compileKotlin")
+    compilerOptions {
+        getOverriddenKotlinLanguageVersion(project)?.let {
+            languageVersion = it
+        }
+        getOverriddenKotlinApiVersion(project)?.let {
+            apiVersion = it
+        }
+        if (isMainTaskName && !unpublished.contains(project.name)) {
+            setWarningsAsErrors(project)
+            freeCompilerArgs.addAll(
+                "-Xexplicit-api=strict",
+                "-Xdont-warn-on-error-suppression",
+            )
+        }
+        configureKotlinUserProject()
+        /* Coroutines do not interop with Java and these flags provide a significant
+         * (i.e. close to double-digit) reduction in both bytecode and optimized dex size */
+        if (this@configureEach is KotlinJvmCompile) {
+            freeCompilerArgs.addAll(
+                "-Xno-param-assertions",
+                "-Xno-call-assertions",
+                "-Xno-receiver-assertions",
+            )
+        }
+        if (this@configureEach is KotlinNativeCompile) {
+            optIn.addAll(
+                "kotlinx.cinterop.ExperimentalForeignApi",
+                "kotlinx.cinterop.UnsafeNumber",
+                "kotlin.experimental.ExperimentalNativeApi",
+                "kotlin.native.concurrent.ObsoleteWorkersApi",
+            )
+        }
+        addExtraCompilerFlags(project)
+    }
 }
